Logical move by the transport-layer to client
